Don't attempt to force the role to a state
Previously, when we started up we would check whether a role was
in the state we expected it to be (ready-to-start or running) and
start or stop it if it didn't match our expectation. This is the
wrong behavior; we need to be reporting the real state and not
arbitrarily altering the system. (For example, someone may have
chosen to stop one of the services through systemd for a good
reason).