Domain Controller: Don't require both forwarders
When specifying forwarders for the domain controller, we were checking
that both the IPv4 and IPv6 forwarders are present, but we actually
only strictly need one or the other (and many environments don't even
have both). Change it so that we check for at least one instead.

Fixes https://github.com/libre-server/rolekit/issues/64